Refactor sample client for Siddhi 5
