+++ modules/siddhi-launcher/bin/siddhi
@@ -181,4 +181,4 @@ $JAVACMD \
 	-Denable.nonblocking=false \
 	-Djava.security.egd=file:/dev/./urandom \
 	-Dfile.encoding=UTF8 \
-	org.wso2.siddhi.sdk.launcher.Main "$@"
+	io.siddhi.sdk.launcher.Main "$@"
