+++ modules/siddhi-launcher/conf/log4j.properties
@@ -23,7 +23,7 @@
 
 log4j.rootLogger=DEBUG, console, Default
 
-log4j.logger.org.wso2=INFO
+log4j.logger.io.siddhi=INFO
 log4j.logger.io.netty=OFF
 
 #File apender
