+++ modules/siddhi-launcher/samples/sample-clients/tcp-client/build.xml
@@ -23,7 +23,7 @@
     <property name="resources.dir" value="src/main/resources"/>
     <property name="temp.dir" value="temp"/>
     <property name="class.dir" value="${temp.dir}/classes"/>
-    <property name="main-class" value="org.wso2.sp.tcp.client.TCPClient"/>
+    <property name="main-class" value="io.siddhi.sdk.samples.tcp.client.TCPClient"/>
     <property name="lib.dir" value="../lib"/>
 
     <property name="host" value="localhost"/>
@@ -71,7 +71,7 @@
 
     <target name="storeTest" depends="compile">
       <echo>Running RDBMS Store Client</echo>
-      <java classname="org.wso2.sp.tcp.client.StoreTestClient"
+      <java classname="io.siddhi.sdk.samples.tcp.client.StoreTestClient"
             classpathref="javac.classpath" fork="true">
           <arg value="${host}"/>
           <arg value="${port}"/>
