--- modules/siddhi-launcher/samples/sample-clients/wso2event-client/src/main/resources/data.agent.config.yaml
@@ -1,131 +1,0 @@
-################################################################################
-#   Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved
-#
-#   Licensed under the Apache License, Version 2.0 (the \"License\");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an \"AS IS\" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-################################################################################
-
-  # Configuration of the Data Agents - to publish events through databridge
-data.agent.config:
-    # Data agent configurations
-    # THIS IS A MANDATORY FIELD
-  agents:
-  - 
-      # Data agent configuration
-    agentConfiguration:
-        # Data agent name
-        # THIS IS A MANDATORY FIELD
-      name: Thrift
-        # Data endpoint class
-        # THIS IS A MANDATORY FIELD
-      dataEndpointClass: org.wso2.carbon.databridge.agent.endpoint.thrift.ThriftDataEndpoint
-        # Data publisher strategy
-      publishingStrategy: async
-        # Trust store path
-      trustStorePath: ''
-        # Trust store password
-      trustStorePassword: ''
-        # Queue Size
-      queueSize: 32768
-        # Batch Size
-      batchSize: 200
-        # Core pool size
-      corePoolSize: 5
-        # Socket timeout in milliseconds
-      socketTimeoutMS: 30000
-        # Maximum pool size
-      maxPoolSize: 10
-        # Keep alive time in pool
-      keepAliveTimeInPool: 20
-        # Reconnection interval
-      reconnectionInterval: 30
-        # Max transport pool size
-      maxTransportPoolSize: 250
-        # Max idle connections
-      maxIdleConnections: 250
-        # Eviction time interval
-      evictionTimePeriod: 5500
-        # Min idle time in pool
-      minIdleTimeInPool: 5000
-        # Secure max transport pool size
-      secureMaxTransportPoolSize: 250
-        # Secure max idle connections
-      secureMaxIdleConnections: 250
-        # secure eviction time period
-      secureEvictionTimePeriod: 5500
-        # Secure min idle time in pool
-      secureMinIdleTimeInPool: 5000
-        # SSL enabled protocols
-      sslEnabledProtocols: TLSv1,TLSv1.1,TLSv1.2
-        # Ciphers
-      ciphers: SSL_RSA_WITH_RC4_128_MD5,SSL_RSA_WITH_RC4_128_SHA,TLS_RSA_WITH_AES_128_CBC_SHA,TLS_DHE_RSA_WITH_AES_128_CBC_SHA,TLS_DHE_DSS_WITH_AES_128_CBC_SHA,SSL_RSA_WITH_3DES_EDE_CBC_SHA,SSL_DHE_RSA_WITH_3DES_EDE_CBC_SHA,SSL_DHE_DSS_WITH_3DES_EDE_CBC_SHA
-  - 
-      # Data agent configuration
-    agentConfiguration:
-        # Data agent name
-        # THIS IS A MANDATORY FIELD
-      name: Binary
-        # Data endpoint class
-        # THIS IS A MANDATORY FIELD
-      dataEndpointClass: org.wso2.carbon.databridge.agent.endpoint.binary.BinaryDataEndpoint
-        # Data publisher strategy
-      publishingStrategy: async
-        # Trust store path
-      trustStorePath: ''
-        # Trust store password
-      trustStorePassword: ''
-        # Queue Size
-      queueSize: 32768
-        # Batch Size
-      batchSize: 200
-        # Core pool size
-      corePoolSize: 5
-        # Socket timeout in milliseconds
-      socketTimeoutMS: 30000
-        # Maximum pool size
-      maxPoolSize: 10
-        # Keep alive time in pool
-      keepAliveTimeInPool: 20
-        # Reconnection interval
-      reconnectionInterval: 30
-        # Max transport pool size
-      maxTransportPoolSize: 250
-        # Max idle connections
-      maxIdleConnections: 250
-        # Eviction time interval
-      evictionTimePeriod: 5500
-        # Min idle time in pool
-      minIdleTimeInPool: 5000
-        # Secure max transport pool size
-      secureMaxTransportPoolSize: 250
-        # Secure max idle connections
-      secureMaxIdleConnections: 250
-        # secure eviction time period
-      secureEvictionTimePeriod: 5500
-        # Secure min idle time in pool
-      secureMinIdleTimeInPool: 5000
-        # SSL enabled protocols
-      sslEnabledProtocols: TLSv1,TLSv1.1,TLSv1.2
-        # Ciphers
-      ciphers: SSL_RSA_WITH_RC4_128_MD5,SSL_RSA_WITH_RC4_128_SHA,TLS_RSA_WITH_AES_128_CBC_SHA,TLS_DHE_RSA_WITH_AES_128_CBC_SHA,TLS_DHE_DSS_WITH_AES_128_CBC_SHA,SSL_RSA_WITH_3DES_EDE_CBC_SHA,SSL_DHE_RSA_WITH_3DES_EDE_CBC_SHA,SSL_DHE_DSS_WITH_3DES_EDE_CBC_SHA
-
-wso2.securevault:
-  secretRepository:
-    type: org.wso2.carbon.secvault.repository.DefaultSecretRepository
-    parameters:
-      privateKeyAlias: wso2carbon
-      keystoreLocation: src/main/resources/securevault.jks
-      secretPropertiesFile: src/main/resources/secrets.properties
-  masterKeyReader:
-    type: org.wso2.carbon.secvault.reader.DefaultMasterKeyReader
-    parameters:
-      masterKeyReaderFile: src/main/resources/master-keys.yaml
