--- modules/siddhi-launcher/samples/sample-clients/wso2event-server/build.xml
@@ -1,84 +1,0 @@
-<!--
-  ~ Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-
-<project default="run">
-
-    <property name="siddhi.home" value="../../../"/>
-    <property name="src.dir" value="src/main/java"/>
-    <property name="resources.dir" value="src/main/resources"/>
-    <property name="temp.dir" value="temp"/>
-    <property name="class.dir" value="${temp.dir}/classes"/>
-    <property name="main-class" value="org.wso2.carbon.sample.server.DatabridgeTestServer"/>
-
-    <property name="host" value="localhost"/>
-    <property name="port" value="7612"/>
-    <property name="protocol" value="thrift"/>
-
-    <path id="javac.classpath">
-        <pathelement path="${class.dir}"/>
-        <fileset dir="${siddhi.home}/lib">
-            <include name="org.apache.commons.logging_*.jar"/>
-            <include name="com.google.gson_*.jar"/>
-            <include name="com.google.guava_*.jar"/>
-            <include name="libthrift_0.9.*.jar"/>
-            <include name="slf4j.log4j12_1.7.*.jar"/>
-            <include name="slf4j.api_*.jar"/>
-            <include name="json_*.jar"/>
-            <include name="org.wso2.carbon.config_2.0.*.jar"/>
-            <include name="org.wso2.carbon.secvault_5.0.*.jar"/>
-            <include name="org.wso2.carbon.utils_2.0.*.jar"/>
-            <include name="org.wso2.carbon.databridge.commons.thrift_6.0.*.jar"/>
-            <include name="org.wso2.carbon.databridge.commons.binary_6.0.*.jar"/>
-            <include name="org.wso2.carbon.databridge.commons_6.0.*.jar"/>
-            <include name="org.wso2.carbon.databridge.core_6.0.*.jar"/>
-            <include name="org.wso2.carbon.databridge.receiver.thrift_6.0.*.jar"/>
-            <include name="org.wso2.carbon.databridge.receiver.binary_6.0.*.jar"/>
-            <include name="org.ops4j.pax.logging.pax-logging-log4j2_*.jar"/>
-            <include name="org.ops4j.pax.logging.pax-logging-api_*.jar"/>
-            <include name="org.yaml.snakeyaml_*.jar"/>
-        </fileset>
-    </path>
-
-    <target name="clean">
-        <delete dir="${class.dir}" quiet="true"/>
-        <delete dir="${temp.dir}"/>
-    </target>
-
-    <target name="init">
-        <mkdir dir="${temp.dir}"/>
-        <mkdir dir="${class.dir}"/>
-    </target>
-
-    <target name="compile" depends="init">
-        <javac srcdir="${src.dir}" destdir="${class.dir}" compiler="modern">
-            <include name="*/**"/>
-            <classpath refid="javac.classpath"/>
-        </javac>
-    </target>
-
-    <target name="run" depends="compile">
-        <echo>Configure -Dhost and -Dport and -Dprotocol</echo>
-        <java classname="${main-class}"
-              classpathref="javac.classpath" fork="true">
-            <arg value="${host}"/>
-            <arg value="${port}"/>
-            <arg value="${protocol}"/>
-        </java>
-    </target>
-</project>
