--- modules/siddhi-launcher/samples/sample-clients/wso2event-server/src/main/java/org/wso2/carbon/sample/server/WSO2EventServerUtil.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.sample.server;
-
-import java.io.File;
-
-/**
- * Util class which sets Keystore and Databridge configuration
- */
-public class WSO2EventServerUtil {
-
-    public static void setKeyStoreParams() {
-        File filePath = new File("src" + File.separator + "main" + File.separator + "resources");
-        if (!filePath.exists()) {
-            filePath = new File("modules" + File.separator + "samples" + File.separator +
-                    "sample-clients" + File.separator + "wso2event-server" + File.separator + "src" +
-                    File.separator + "main" + File.separator + "resources");
-        }
-        if (!filePath.exists()) {
-            filePath = new File("resources");
-        }
-        if (!filePath.exists()) {
-            filePath = new File("test" + File.separator + "resources");
-        }
-        String keyStore = filePath.getAbsolutePath();
-        System.setProperty("Security.KeyStore.Location", keyStore + File.separator + "wso2carbon.jks");
-        System.setProperty("Security.KeyStore.Password", "wso2carbon");
-    }
-
-    public static String getDataBridgeConfigPath() {
-        File filePath = new File("src" + File.separator + "main" + File.separator + "resources");
-        if (!filePath.exists()) {
-            filePath = new File("modules" + File.separator + "samples" + File.separator +
-                    "sample-clients" + File.separator + "wso2event-server" + File.separator + "src" +
-                    File.separator + "main" + File.separator + "resources");
-        }
-        if (!filePath.exists()) {
-            filePath = new File("resources");
-        }
-        if (!filePath.exists()) {
-            filePath = new File("test" + File.separator + "resources");
-        }
-        return filePath.getAbsolutePath() + File.separator + "databridge.config.yaml";
-    }
-
-}
