--- modules/siddhi-launcher/samples/sample-clients/wso2event-server/src/main/resources/log4j.properties
@@ -1,34 +1,0 @@
-#
-# Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-
-# For the general syntax of property based configuration files see the
-# documenation of org.apache.log4j.PropertyConfigurator.
-
-# The root category uses the appender called A1. Since no priority is
-# specified, the root category assumes the default priority for root
-# which is DEBUG in log4j. The root category is the only category that
-# has a default priority. All other categories need not be assigned a
-# priority in which case they inherit their priority from the
-# hierarchy.
-
-#log4j.rootLogger=debug, console
-log4j.rootLogger=info, console
-
-log4j.appender.console=org.apache.log4j.ConsoleAppender
-log4j.appender.console.layout=org.apache.log4j.PatternLayout
-#log4j.appender.console.layout.ConversionPattern=[%t] %-5p %c %x - %m%n
-log4j.appender.console.layout.ConversionPattern=[%t] %-5p %c %x - %m%n
