--- modules/siddhi-launcher/samples/sample-clients/wso2event-server/src/main/resources/master-keys.yaml
@@ -1,6 +1,0 @@
-permanent: true
-# mastterKeys should be given in the base64 format and the data type of the
-# password should be explicitly specified as binary (type !!binary).
-masterKeys:
-  keyStorePassword: !!binary d3NvMmNhcmJvbg==
-  privateKeyPassword: !!binary d3NvMmNhcmJvbg==
