--- modules/siddhi-launcher/samples/sample-clients/wso2event-server/src/main/resources/secrets.properties
@@ -1,2 +1,0 @@
-wso2.sample.password1=plainText ABC@123
-wso2.sample.password2=cipherText o4NUR97O+zBXuqkVy+hjnOeCPta+QNnqRTsqf1TPT1DagtmI+ho6rfD9AMC5Ud0xstks7KOq/hKVUnc3LTa25To8zAzWUSCeccbR5A16OGQ7DXFvEZgDhsC1fBCmKncPVLFE8oHcr6K76dWCJZhQa+SStCLY+YCDNuJC4aTKhCfbU3y817mn+ozXTBG3Sac4PO/0cBDctGkd9K6BOWm4tPHpzvEmz4OpQEoEaXpoJXaiBZXTN1/LOiwN5croVBIwp9fju4haN54ekMP8/adAsKVpomSfOVLEKMPQdDj+G0sDmKZKw+uGXzCJ2I4gxF2nzCIAvH9VrNfZZBZ9ZuaGQg\=\=
